LayoutTests/canvas/philip/tests/2d.imageData.put.nonfinite-expected.txt

< [index] >
2d.imageData.put.nonfinite
putImageData() throws NOT_SUPPORTED_ERR if arguments are not finite
References: 2d.imageData.put.nonfinite
Actual output:
Passed
